+++ database/mongodb/.env
@@ -1,3 +1,3 @@
-DATABASE_URL=
+YOUR_RESOURCE_PRIMARY_CONNECTION_STRING=
 DATABASE_NAME=BooksDB
 DATABASE_COLLECTION_NAME=GoodBooks
